# api/openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio API
  description: Backend for Portfolio - Migrated from TS to Go
  version: 1.1.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /auth/login:
    post:
      summary: Admin login
      operationId: postLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # ─── Projects ───
  /projects:
    get:
      summary: Get all portfolio projects
      operationId: getProjects
      responses:
        '200':
          description: A list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Create a new project
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/reorder:
    put:
      summary: Reorder projects
      operationId: reorderProjects
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ReorderItem'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /projects/{id}:
    put:
      summary: Update a project
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      summary: Delete a project
      operationId: deleteProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # ─── Experience ───
  /experience:
    get:
      summary: Get work experience list
      operationId: getExperience
      responses:
        '200':
          description: A list of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Experience'
    post:
      summary: Create a new experience entry
      operationId: createExperience
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Experience'
      responses:
        '201':
          description: Experience created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experience'
  /experience/reorder:
    put:
      summary: Reorder experience entries
      operationId: reorderExperience
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ReorderItem'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /experience/{id}:
    put:
      summary: Update an experience entry
      operationId: updateExperience
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Experience'
      responses:
        '200':
          description: Experience updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experience'
    delete:
      summary: Delete an experience entry
      operationId: deleteExperience
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Experience deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # ─── Skills ───
  /skills:
    get:
      summary: Get skills categorized
      operationId: getSkills
      responses:
        '200':
          description: A list of skill categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillCategory'
    post:
      summary: Create a new skill category
      operationId: createSkillCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillCategory'
      responses:
        '201':
          description: Skill category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillCategory'
  /skills/reorder:
    put:
      summary: Reorder skill categories
      operationId: reorderSkills
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ReorderItem'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /skills/{id}:
    put:
      summary: Update a skill category
      operationId: updateSkillCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillCategory'
      responses:
        '200':
          description: Skill category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillCategory'
    delete:
      summary: Delete a skill category
      operationId: deleteSkillCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Skill category deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # ─── Contact ───
  /contact:
    get:
      summary: List contact messages (admin)
      operationId: getContactMessages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactRequest'
    post:
      summary: Submit contact form
      operationId: postContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactRequest'
      responses:
        '201':
          description: Message received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Parameters ───
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectID as string
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      required: [token]
      properties:
        token: { type: string }

    # ─── Projects ───
    Project:
      type: object
      required: [_id, slug, title, summary, stack, links]
      properties:
        _id:
          type: string
          example: "65b2f1234567890abcdef123"
        slug: { type: string }
        title: { type: string }
        summary: { type: string }
        details: { type: array, items: { type: string } }
        stack: { type: array, items: { type: string } }
        links:
          type: object
          properties:
            demo: { type: string }
            repo: { type: string }
        image: { type: string }
        contributors: { type: string }
        order: { type: integer }

    # ─── Experience ───
    Experience:
      type: object
      required: [_id, role, company, start]
      properties:
        _id: { type: string }
        role: { type: string }
        company: { type: string }
        url: { type: string }
        start: { type: string }
        end: { type: string }
        location: { type: string }
        tags: { type: array, items: { type: string } }
        description: { type: array, items: { type: string } }
        order: { type: integer }

    # ─── Skills ───
    SkillItem:
      type: object
      required: [name, level]
      properties:
        name: { type: string }
        level: { type: string }
        icon: { type: string }
        description: { type: string }
    SkillCategory:
      type: object
      required: [_id, title]
      properties:
        _id: { type: string }
        title: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/SkillItem' } }
        order: { type: integer }

    # ─── Reorder ───
    ReorderItem:
      type: object
      required: [id, order]
      properties:
        id: { type: string }
        order: { type: integer }

    # ─── Auth User ───
...
    ContactRequest:
      type: object
      required: [name, email, message]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        subject: { type: string }
        message: { type: string }

    # ─── Standard Response ───
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }